audit event refactor
